feat: coordinate STT sessions on native clients
